get flir video file creation ready to run as end-of-day script in production
